requirements.txt
crewai>=0.28.8
crewai-tools>=0.1.6
langchain>=0.1.10
langchain-community>=0.0.20
langchain-google-genai>=1.0.1
# langchain-openai>=0.0.5
# langchain-anthropic>=0.1.0
google-generativeai>=0.3.0
# langchain-ollama>=0.1.0
langchain-huggingface>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
pandas>=2.1.4
numpy>=1.24.3
matplotlib>=3.8.2
seaborn>=0.13.0
litellm>=1.74.9